Implement viewModel factory
