+++ fearless/Modules/Staking/StakingRewardPayouts/StakingRewardPayoutsPresenter.swift
@@ -7,98 +7,14 @@ final class StakingRewardPayoutsPresenter {
     var wireframe: StakingRewardPayoutsWireframeProtocol!
     var interactor: StakingRewardPayoutsInteractorInputProtocol!
 
-    private let addressFactory = SS58AddressFactory()
-    private let chain: Chain
-    private let balanceViewModelFactory: BalanceViewModelFactoryProtocol
-    private lazy var formatterFactory = AmountFormatterFactory()
     private var payoutsInfo: PayoutsInfo?
     private var priceData: PriceData?
+    private let chain: Chain
+    private let viewModelFactory: StakingPayoutViewModelFactoryProtocol
 
-    init(
-        chain: Chain,
-        balanceViewModelFactory: BalanceViewModelFactoryProtocol
-    ) {
+    init(chain: Chain, viewModelFactory: StakingPayoutViewModelFactoryProtocol) {
         self.chain = chain
-        self.balanceViewModelFactory = balanceViewModelFactory
-    }
-
-    private func createCellViewModels(
-        for payoutsInfo: PayoutsInfo
-    ) -> [StakingRewardHistoryCellViewModel] {
-        payoutsInfo.payouts.map { payout in
-            let daysLeftText = daysLeftAttributedString(
-                activeEra: payoutsInfo.activeEra,
-                payoutEra: payout.era,
-                historyDepth: payoutsInfo.historyDepth
-            )
-
-            return StakingRewardHistoryCellViewModel(
-                addressOrName: self.addressTitle(payout),
-                daysLeftText: daysLeftText,
-                tokenAmountText: "+" + self.tokenAmountText(payout.reward),
-                usdAmountText: priceText(payout.reward, priceData: priceData)
-            )
-        }
-    }
-
-    private func addressTitle(_ payout: PayoutInfo) -> String {
-        if let displayName = payout.identity?.displayName {
-            return displayName
-        }
-
-        if let address = try? addressFactory
-            .addressFromAccountId(data: payout.validator, type: chain.addressType) {
-            return address
-        }
-
-        return ""
-    }
-
-    private func tokenAmountText(_ value: Decimal) -> String {
-        let locale = view?.localizationManager?.selectedLocale ?? Locale.current
-        return balanceViewModelFactory.amountFromValue(value).value(for: locale)
-    }
-
-    private func priceText(_ amount: Decimal, priceData: PriceData?) -> String {
-        guard let priceData = priceData else {
-            return "$0"
-        }
-
-        let locale = view?.localizationManager?.selectedLocale ?? Locale.current
-        let price = balanceViewModelFactory
-            .priceFromAmount(amount, priceData: priceData).value(for: locale)
-        return price
-    }
-
-    private func daysLeftAttributedString(
-        activeEra: EraIndex,
-        payoutEra: EraIndex,
-        historyDepth: UInt32
-    ) -> NSAttributedString {
-        let eraDistance = historyDepth - (activeEra - payoutEra)
-        let daysLeft = eraDistance / UInt32(chain.erasPerDay)
-        let daysLeftText = daysLeft == 1 ? " day left" : " days left"
-
-        let historyDepthDays = (historyDepth / 2) / UInt32(chain.erasPerDay)
-        let textColor: UIColor = daysLeft < historyDepthDays ?
-            R.color.colorRed()! : R.color.colorLightGray()!
-
-        let attrubutedString = NSAttributedString(
-            string: daysLeft.description + daysLeftText,
-            attributes: [.foregroundColor: textColor]
-        )
-        return attrubutedString
-    }
-
-    private func defineBottomButtonTitle(
-        for payouts: [PayoutInfo]
-    ) -> String {
-        let totalReward = payouts
-            .reduce(into: Decimal(0)) { reward, payout in
-                reward += payout.reward
-            }
-        let amountText = tokenAmountText(totalReward)
-        return "Payout all (\(amountText))"
+        self.viewModelFactory = viewModelFactory
     }
 
     private func updateView() {
@@ -111,10 +27,7 @@ final class StakingRewardPayoutsPresenter {
             return
         }
 
-        let viewModel = StakingPayoutViewModel(
-            cellViewModels: createCellViewModels(for: payoutsInfo),
-            bottomButtonTitle: defineBottomButtonTitle(for: payoutsInfo.payouts)
-        )
+        let viewModel = viewModelFactory.createPayoutsViewModel(payoutsInfo: payoutsInfo, priceData: priceData)
         let viewState = StakingRewardPayoutsViewState.payoutsList(viewModel)
         view?.reload(with: viewState)
     }
